common: Move fi_nic code to common one
This allows to use ofi_nic_* internal API by providers that are
built with dl option. This fixes usNIC provider that is already
using the ofi_nic API.
