+++ include/ofi.h
@@ -40,6 +40,7 @@
 #include <pthread.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdarg.h>
 #include <sys/param.h>
 #include <netinet/in.h>
 #include <ifaddrs.h>
@@ -85,6 +86,19 @@ extern "C" {
 
 #define OFI_MAGIC_64 (0x0F1C0DE0F1C0DE64)
 
+#define TAB "    "
+
+#define CASEENUMSTR(SYM) \
+	case SYM: { ofi_strcatf(buf, #SYM); break; }
+#define IFFLAGSTR(flags, SYM) \
+	do { if (flags & SYM) ofi_strcatf(buf, #SYM ", "); } while(0)
+#define CASEENUMSTRN(SYM, N) \
+	case SYM: { ofi_strncatf(buf, N, #SYM); break; }
+#define IFFLAGSTRN(flags, SYM, N) \
+	do { if (flags & SYM) ofi_strncatf(buf, N, #SYM ", "); } while(0)
+
+#define ofi_strcatf(dest, ...) \
+	ofi_strncatf(dest, OFI_BUFSIZ, __VA_ARGS__)
 
 /*
  * CPU specific features
@@ -140,6 +154,8 @@ void ofi_hook_init(void);
 void ofi_hook_fini(void);
 void ofi_hook_install(struct fid_fabric *hfabric, struct fid_fabric **fabric,
 		      struct fi_provider *prov);
+void ofi_remove_comma(char *buffer);
+void ofi_strncatf(char *dest, size_t n, const char *fmt, ...);
 
 const char *ofi_hex_str(const uint8_t *data, size_t len);
