+++ src/fabric.c
@@ -543,155 +543,6 @@ FI_DESTRUCTOR(fi_fini(void))
 	ofi_osd_fini();
 }
 
-/* The provider must free any prov_attr data prior to calling this
- * routine.
- */
-int ofi_nic_close(struct fid *fid)
-{
-	struct fid_nic *nic = (struct fid_nic *) fid;
-
-	assert(fid && fid->fclass == FI_CLASS_NIC);
-
-	if (nic->device_attr) {
-		free(nic->device_attr->name);
-		free(nic->device_attr->device_id);
-		free(nic->device_attr->device_version);
-		free(nic->device_attr->vendor_id);
-		free(nic->device_attr->driver);
-		free(nic->device_attr->firmware);
-		free(nic->device_attr);
-	}
-
-	free(nic->bus_attr);
-
-	if (nic->link_attr) {
-		free(nic->link_attr->address);
-		free(nic->link_attr->network_type);
-		free(nic->link_attr);
-	}
-
-	free(nic);
-	return 0;
-}
-
-int ofi_nic_control(struct fid *fid, int command, void *arg)
-{
-	struct fid_nic *nic = container_of(fid, struct fid_nic, fid);
-	struct fid_nic **dup = (struct fid_nic **) arg;
-
-	switch(command) {
-	case FI_DUP:
-		*dup = ofi_nic_dup(nic);
-		return *dup ? FI_SUCCESS : -FI_ENOMEM;
-	default:
-		return -FI_ENOSYS;
-	}
-}
-
-struct fi_ops default_nic_ops = {
-	.size = sizeof(struct fi_ops),
-	.close = ofi_nic_close,
-	.control = ofi_nic_control,
-	.tostr = ofi_nic_tostr,
-};
-
-static int ofi_dup_dev_attr(const struct fi_device_attr *attr,
-			    struct fi_device_attr **dup_attr)
-{
-	*dup_attr = calloc(1, sizeof(**dup_attr));
-	if (!*dup_attr)
-		return -FI_ENOMEM;
-
-	if (ofi_str_dup(attr->name, &(*dup_attr)->name) ||
-	    ofi_str_dup(attr->device_id, &(*dup_attr)->device_id) ||
-	    ofi_str_dup(attr->device_version, &(*dup_attr)->device_version) ||
-	    ofi_str_dup(attr->vendor_id, &(*dup_attr)->vendor_id) ||
-	    ofi_str_dup(attr->driver, &(*dup_attr)->driver) ||
-	    ofi_str_dup(attr->firmware, &(*dup_attr)->firmware))
-		return -FI_ENOMEM;
-
-	return 0;
-}
-
-static int ofi_dup_bus_attr(const struct fi_bus_attr *attr,
-			    struct fi_bus_attr **dup_attr)
-{
-	*dup_attr = calloc(1, sizeof(**dup_attr));
-	if (!*dup_attr)
-		return -FI_ENOMEM;
-
-	**dup_attr = *attr;
-	return 0;
-}
-
-static int ofi_dup_link_attr(const struct fi_link_attr *attr,
-			     struct fi_link_attr **dup_attr)
-{
-	*dup_attr = calloc(1, sizeof(**dup_attr));
-	if (!*dup_attr)
-		return -FI_ENOMEM;
-
-	if (ofi_str_dup(attr->address, &(*dup_attr)->address) ||
-	    ofi_str_dup(attr->network_type, &(*dup_attr)->network_type))
-		return -FI_ENOMEM;
-
-	(*dup_attr)->mtu = attr->mtu;
-	(*dup_attr)->speed = attr->speed;
-	(*dup_attr)->state = attr->state;
-	return 0;
-}
-
-struct fid_nic *ofi_nic_dup(const struct fid_nic *nic)
-{
-	struct fid_nic *dup_nic;
-	int ret;
-
-	dup_nic = calloc(1, sizeof(*dup_nic));
-	if (!dup_nic)
-		return NULL;
-
-	if (!nic) {
-		dup_nic->fid.fclass = FI_CLASS_NIC;
-		dup_nic->device_attr = calloc(1, sizeof(*dup_nic->device_attr));
-		dup_nic->bus_attr = calloc(1, sizeof(*dup_nic->bus_attr));
-		dup_nic->link_attr = calloc(1, sizeof(*dup_nic->link_attr));
-
-		if (!dup_nic->device_attr || !dup_nic->bus_attr ||
-		    !dup_nic->link_attr)
-			goto fail;
-
-		dup_nic->fid.ops = &default_nic_ops;
-		return dup_nic;
-	}
-
-	assert(nic->fid.fclass == FI_CLASS_NIC);
-	dup_nic->fid = nic->fid;
-
-	if (nic->device_attr) {
-		ret = ofi_dup_dev_attr(nic->device_attr, &dup_nic->device_attr);
-		if (ret)
-			goto fail;
-	}
-
-	if (nic->bus_attr) {
-		ret = ofi_dup_bus_attr(nic->bus_attr, &dup_nic->bus_attr);
-		if (ret)
-			goto fail;
-	}
-
-	if (nic->link_attr) {
-		ret = ofi_dup_link_attr(nic->link_attr, &dup_nic->link_attr);
-		if (ret)
-			goto fail;
-	}
-
-	return dup_nic;
-
-fail:
-	ofi_nic_close(&dup_nic->fid);
-	return NULL;
-}
-
 __attribute__((visibility ("default"),EXTERNALLY_VISIBLE))
 void DEFAULT_SYMVER_PRE(fi_freeinfo)(struct fi_info *info)
 {
